+++ simplemma/lemmatizer.py
